pyproject.toml
[project]
name = "agora-orchestrator"
version = "0.1.0"
description = "Local-first Reddit + Ollama + Telegram sales outreach orchestrator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "praw>=7.8.1",
  "python-telegram-bot[job-queue]>=22.0",
]

[project.scripts]
agora = "orchestrator.main:main"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
